chore(slug): update slug form stories to sb v7
